refactor: move CLI-related functions to `internals/cli` package
