+++ internals/cli/cli_flags.go
@@ -1,10 +1,6 @@
-package main
+package cli
 
-import (
-	"flag"
-
-	"dougdomingos.com/image-filters/filters"
-)
+import "flag"
 
 // ParsedFlags declares a structure used for retrieving the flag values
 // provided by the user.
@@ -16,33 +12,22 @@ type ParsedFlags struct {
 	ListPipelines bool
 }
 
-// parseInputFlags declares the expected command-line flags and parses their
+// ParseInputFlags declares the expected command-line flags and parses their
 // values into a ParsedFlags struct for later usage.
-func parseInputFlags() ParsedFlags {
+func ParseInputFlags() ParsedFlags {
 	imgPath := flag.String("img", "", "Path to image file")
 	outputDir := flag.String("outDir", "./output", "Directory where the processed image shall be stored")
 	filterName := flag.String("filter", "", "Name of the filter pipeline to be applied")
 	isConcurrent := flag.Bool("concurrent", false, "Specify if filter pipeline should be use parallel execution")
-	listPipelines := flag.Bool("list", false, "List all avaliable filter pipelines")
+	listPipelines := flag.Bool("listPipelines", false, "Specify if filter pipeline should be use parallel execution")
+
 	flag.Parse()
 
 	return ParsedFlags{
-		ImgPath:       *imgPath,
-		OutputDir:     *outputDir,
-		FilterName:    *filterName,
-		IsConcurrent:  *isConcurrent,
+		ImgPath:      *imgPath,
+		OutputDir:    *outputDir,
+		FilterName:   *filterName,
+		IsConcurrent: *isConcurrent,
 		ListPipelines: *listPipelines,
 	}
-}
-
-// getAvaliablePipelines returns the string identifiers of all avaliable filter
-// pipelines.
-func getAvaliablePipelines() []string {
-	keys := make([]string, 0, len(filters.AvaliableFilters))
-
-	for filterKey := range filters.AvaliableFilters {
-		keys = append(keys, filterKey)
-	}
-
-	return keys
 }
